Implementacion de addMedication scrern para cargar medicamentos

+++ app/src/main/java/ar/ort/edu/proyecto_final_grupo_4/data/dao/ScheduleDao.kt
@@ -13,7 +13,7 @@ interface ScheduleDao {
     suspend fun insertSchedule(schedule: Schedule)
 
     @Query("SELECT * FROM schedule WHERE medicationID = :medicationId")
-    suspend fun getSchedulesForMedication(medicationId: Int): List<Schedule>
+    suspend fun getSchedulesForMedication(medicationId: Long): List<Schedule>
 
     @Query("SELECT * FROM schedule WHERE scheduleID = :id")
     suspend fun getById(id: Int): Schedule?
